:recycle: [Refactor]: 북마크 컴포넌트 분리
